Porting graphics over to a scenegraph basis.
Significant, although fairly straightforward, changes
to the infrastructure under the declarative rendering
api that spork.sketch uses.  In an effort to make the
rendering more portable, namely to provide the ability
to render to other targets rather than a graphics2d
canvas, defined some convenience macros around the
fundamental scenegraph nodes: translation, scale, etc.
We now generate maps that identify the type of node
and its children, more like svg and other scene graph
formats.  For convenenience, shape implementations are
preserved, but packed into meta data, and the IShape
protocol extends over maps now.  So maps created with
the appropriate information, or maps having :children
keys that are IShape compliant are now directly
renderable as shapes, and have bounds.

The upside of this is that we can do 2 big things:
translate the nested set of maps and vectors (groups of
shapes) into other data structures fairly easily.
The first goal is to flesh out the piccolo2d interpreter
so that we can declaratively build scenegraphs from
rather than imperatively smashing them together (ugh).

Second, we can tag elements in the scene graph as
part of our declaration (much like SVG has properties
for nodes).  This opens up the ability for us to template
and to provide filtering/querying functionality.  Since
Piccolo2d supports properties on nodes, the correspondence
with metadata should map nicely and preserve and
identifiers we provide as part of the scene specification.
This should also map over nicely to HTML5 or js
scenegraph libs, and even SVG export.