finished table updates.
Added new table->lines function, which provides a
seqable and reducible collection of delimite lines.
tables are now optionally primitive-backed, based on
the options passed to tabdelimited->table (specifically
if a schema is passed, we enforce it and error if it's
different than expected, i.e. missing fields).  Future
change could have tabdelimited->table automatically
derive the schema based on the first line, if no
schema is provided, to speed up parsing.)