Added data.sparse for sparse columns, vectors.
Created this in response to wanting a light
weight view of in-memory tables that are
column-backed. Also extended a preliminary
set of functions for getting sequences of
flyrecords out of tables.  Thinking of
changing them to reducible/foldable instances
and using flyrecords as the default instead
of naive record sequences.  need to verify
and test performance implications.  Also,
we may not need to keep the table in memory,
at which point the column-store is less
useful than a row-store, which may serve as
a better logical unit of organziation, i.e.
more amenable to filtering and dropping records,
and allowing garbage collection (at the cost of
having to filter more).