Significant performance updates.
Working on the simulation framework, specifically
allowing the step-function to mutate in place.
Implemented a mutable arraymap that conforms to
clojure's interfaces for persistent structures
(tsk tsk...), as well as performant wrappers for
mutable maps and collections.  Most of this
obviates the work I was doing on the universal
assoc/update functionality in spork.util.collections
, however there's still some useful functionality in
there that could be merged.  data.mutable now
contains several useful datastructures that can
be used to efficiently bash in place without paying
much cost.  The arraymap is pretty damn efficient, with
the caveat being (currently) identity-based semantics.
So, if you're storing keys or strings, identical?
works fine and fast.  Outside of that use case, where
we really need eq / equiv, the associative structures
won't work properly.  I'm banking on that not being
the case though, at least for my needs (speeding up
a portion of a computation using beneficial side
effects).