Migrated to protocols and other performance impro
Working on refactoring/simplifying the entity
messaging system in ai.testing, coalescing on
a performant solution - it seems - that
I should be able to to into or supplant
the original sim infrastructure.

Entities now can participate in a behavioral
context, in which we the entity is updated
according to its unique behavior - or a
default one.  The updates are currently
weakly persistent (meaning that the
behavior environment is using atoms
rather than a mutabl core), but we can
switch to pure if necessary.  For performance
and simplicity, the current solution is decent.

We're getting pretty good throughput for the
old stress test (~8K entities moving across
a 13 year time horizon).  Additionally, for
smaller-scale populations, on parallel runs,
we get very nice throughput too.  I'd like to
get it down into the ms range, but it'll
require some significant optimization work to
get there.  The current performance profile is
adequate, with the ability to use the existing
abstaction and shift to mutation/async stuff.