fixed wierd abstract method error in entitystore
Specifically, the passmap type was getting a "new"
(gah!) abstract method error, because kv-reduce was
being called on a passmap, and it was defaulting to
iter-reduce since no implementation of kv-reduce
was around.  This went to an empty abstract
method....hence the error.  I think we changed
mergee to use reduce-kv, and it precipated this.
No idea why it took this long to occur.
Tests are passing, though..