update simcontext and entitystore
Added register-routes to allow declarative routing
for simcontext, which is just a wrapper around
spork.sim.pure.network/register-routes (probably
should move this out somewhere else).

Also added an aux function to spork.entitysystem.store,
map-component, which allows the caller to map a function
f across the entries of a component c in an entitystore.
A potential performance optimization would be to push this
into the protocol-level, and let the types provide
an internal implementation.  Currnetly, we're paying
overhead of mapping entry-by-entry, which includes
lookups and assoc costs.