Performance enhancement to sim agenda, added cell.
Turns out the api for simcontext encouraged getting
the current time via the next event frequently,
which was - rather than O(1) op - resulting in
a traversal to the lowest child of a tree map,
O(lg(n)) performance.  Significantly improved
time by caching the result lazily, also added
a new datatype that can be used as an alterative
to mutable references in deftypes, but is
compatible with defrecord.  it's in
spork.data.cell, and basically is a serializable
container.