Added some extra parsers to parsing, passmap issue
There's an abstractmethod error that gets
thrown when you try to get a passmap
lazily joined from an underlying map that
doesn't have the row.  Trying to isolate
and fix.  Causes a secondary, hard
to trace null pointer error if there's
something that operates on the passmap.

So, like (:name (get-entity ...)) could
crap out with an opaque null-pointer error
and no explanation/stack trace.

TRying to avoid!