Patched clipboard for compatibility with substance
Ran into a nasty subtle bug where the system
clipboard logic was upsetting the Substance
look and feel and throwing odd runtime errors.

It turns out, the clipboard class reference
is somehow invalidated after Substance installs
its look and feel.  So the original premise of
storing a static reference to the sole system
clipboard doesn't work.

If we move to a function get-clipboard that
looks up the -current- default toolkit's clipboard
and get a reference, then we appease finnicky
PLAF's like Substance.